+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
-Disable checks with nmap if a proxy is specified
+ * Disable checks with nmap if a proxy is specified
+ * Added   ```--require-x-frame-options``` to check for the X-Frame-Options header